Initial Model
The initial model was the binary classifier and as we saw that we got 98% accuracy on the test data set, we are going to improve our output. So, instead of just classifying that if the room is dirty or not, we are going to come up with a model that can find the dirt/garbage in the room.